Desafio Aventureiro - Desenvolvendo

+++ CartasSuperTrunfo.c
@@ -9,7 +9,7 @@ struct CartaSuperTrunfo {
    char estado[3];            //Sigla do estado
    char codigo[4];            //Codigo da carta
    char nomeCidade[100];      //Nome da cidade
-   int populacao;             //População da cidade
+   unsigned long int populacao;             //População da cidade
    float area;                //Area da cidade
    float pib;                 //PIB
    int pontosTuristicos;      //Numero de pontos turisticos
@@ -35,7 +35,7 @@ int main () {
     scanf (" %[^\n]", carta1.nomeCidade);
 
     printf ("Digite a população: ");
-    scanf ("%d",& carta1.populacao);
+    scanf ("%lu",& carta1.populacao);
 
     printf ("Digite a area (km²): ");
     scanf ("%f", &carta1.area);
@@ -46,8 +46,8 @@ int main () {
     printf ("Digite o número de pontos turísticos: ");
     scanf ("%d", &carta1.pontosTuristicos);
 
-    carta1.denspopulacional = carta1.populacao/carta1.area;
-    carta1.pibpercapita = (carta1.pib * 1000000000) / carta1.populacao;
+    carta1.denspopulacional = (float)carta1.populacao/carta1.area;
+    carta1.pibpercapita = (carta1.pib * 1000000000.0f) / carta1.populacao;
 
     // entrada de dados carta2
 
@@ -63,7 +63,7 @@ int main () {
     scanf (" %[^\n]", carta2.nomeCidade);
 
     printf ("Digite a população: ");
-    scanf ("%d",& carta2.populacao);
+    scanf ("%lu",& carta2.populacao);
 
     printf ("Digite a area (km²): ");
     scanf ("%f", &carta2.area);
@@ -74,15 +74,15 @@ int main () {
     printf ("Digite o número de pontos turísticos: ");
     scanf ("%d", &carta2.pontosTuristicos);
 
-    carta2.denspopulacional = carta2.populacao/carta2.area;
-    carta2.pibpercapita = (carta2.pib * 1000000000) / carta2.populacao;
+    carta2.denspopulacional = (float)carta2.populacao/carta2.area;
+    carta2.pibpercapita = (carta2.pib * 1000000000.0f) / carta2.populacao;
 
     //impressão Carta 1
     printf ("\ncarta 1:\n");
     printf ("Estado: %s\n", carta1.estado);
     printf ("Código: %s\n", carta1.codigo);
     printf ("Nome da cidade: %s\n", carta1.nomeCidade);
-    printf ("População: %d\n", carta1.populacao);
+    printf ("População: %lu\n", carta1.populacao);
     printf ("Área: %.2f km²\n", carta1.area);
     printf ("PIB: %.2f bilhoes de reais\n", carta1.pib);
     printf ("Número de pontos turísticos: %d\n", carta1.pontosTuristicos);
@@ -96,7 +96,7 @@ int main () {
     printf ("Estado: %s\n", carta2.estado);
     printf ("Código: %s\n", carta2.codigo);
     printf ("Nome da cidade: %s\n", carta2.nomeCidade);
-    printf ("População: %d\n", carta2.populacao);
+    printf ("População: %lu\n", carta2.populacao);
     printf ("Área: %.2f km²\n", carta2.area);
     printf ("PIB: %.2f bilhoes de reais\n", carta2.pib);
     printf ("Número de pontos turísticos: %d\n", carta2.pontosTuristicos);
